requirements.txt
flask
flask-cors
deta
python_dotenv
